Add license tag for cowsay-org, stating GPL-3.0 (not -or-later)
Based on what I can see in the classic cowsay's Git history, the GPL re-licensing in 2016 was done with just GPL v3.0, and there was no "or later" statement in a license application document or header. So I'm treating it as probably GPL 3.0 only, but not confident enough in that to use the GPL-3.0-only license identifier (yet), so leaving it as the vaguer "GPL-3.0" for now.

See https://github.com/cowsay-org/cowsay/issues/64 for ongoing investigation.